Use i18n with_locale instead of locale setter
The behaviour between i18n and faktory ruby worker is not the same.
Faktory ruby worker restart the locale to the default when i18n reset
the locale with the current locale.

To avoid any mistake I propose to use with_locale method. This method
wrap the locale reset. This avoid logic duplication and keep the logic
of locale reset in the i18n library.